Syncing my work with repo.
* Note: this is not supposed to work. :D Crashes like a bitch.
* Changed startup order a little.
* Made Host Cache Discovery use the same thread.
* Changed CHostCacheHost to use getters and setters. This allows preventing timestamp changes.
* Optimized find() by permanently storing a linked list Iterator in the CHostCacheHost class. According to the spec, these iterators are only invalidated if the exact same item they point to is removed from the list. Theoretically, the item can even be moved around in the list randomly.
* Changed the Host Cache internal storage. Don't like it however, so this is probably going to change again...
* Used QSet for the getConnectable() exception list.
* Changed settings: Connection.FailureLimit from int to quint8, Gnutella.ConnectThrottle and Gnutella.HostCacheSize from int to quint32.

+++ Quazaa/quazaasettings.h
@@ -162,7 +162,7 @@ namespace Settings
 	{
 		bool		DetectConnectionLoss;					// Detect loss of internet
 		bool		DetectConnectionReset;					// Detect regaining of internet connection
-		int			FailureLimit;							// Max allowed connection failures (default = 3) (Neighbour connections)
+		quint8		FailureLimit;							// Max allowed connection failures (default = 3) (Neighbour connections)
 		int			FailurePenalty;							// Delay after connection failure (seconds, default = 300) (Neighbour connections)
 		quint64		InSpeed;								// Inbound internet connection speed in B/s
 		quint64		OutSpeed;								// Outbound internet connection speed in B/s
@@ -279,9 +279,9 @@ namespace Settings
 	struct sGnutella
 	{
 		int			ConnectFactor;
-		int			ConnectThrottle;						// Delay between connection attempts (seconds)
+		quint32		ConnectThrottle;						// Delay between connection attempts (seconds)
 		int			HitsPerPacket;							// Max hits allowed for a single packet
-		int			HostCacheSize;							// Size of host cache
+		quint32		HostCacheSize;							// Max size of host cache; 0 to disable
 		int			HostCacheView;
 		int			MaxHits;								// Max hits allowed for a single query
 		quint32		MaxResults;								// Maximum results to return to a single query
